Add files via upload
Now all the necessary functions for both the frequency and the asymmetrical coupling sweep are stored in the HeatCurrentFunctions file. All the files containing different results, first import the HeatCurrentFunctions file as QH , then sets up the parameters, finally calculating and plotting the results. As a reference, Marcello's original code was also included here, which also uses HeatCurrentFunctions as well.